requirements.txt
asgiref==3.4.1
astroid==2.6.6
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.1
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.5
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-filter==21.1
django-rest-knox==4.1.0
django-tinymce==3.4.0
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.4.2
Flask==2.0.1
function==1.2.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.3.2
pipreqs==0.4.11
platformdirs==2.4.1
playsound==1.3.0
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.8
virtualenv==20.13.0
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
yarg==0.1.9
zipp==3.6.0
